Automata updates now happen on a separate thread

+++ src/lib/objects/mod.rs
@@ -8,7 +8,7 @@ use crate::Vertex;
 
 pub struct MeshObjectData {
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<u16>
+    pub indices: Vec<u32>
 }
 
 pub trait MeshObject: private::MeshObject {
